requirements.txt
dropbox
pandas
spacy
pythonmemcached
sentence-transformers
faiss-cpu
langchain
openai
python-dotenv